fix beautiful soup warning
UserWarning: No parser was explicitly specified, so I'm using the best
available HTML parser for this system ("lxml"). This usually isn't
a problem, but if you run this code on another system, or in a different
virtual environment, it may use a different parser and behave
differently.

The code that caused this warning is on line 1997 of the file
/var/mirror/eggs/_venv/lib/python2.7/site-packages/flask/app.py. To get
rid of this warning, change code that looks like this:

 BeautifulSoup(YOUR_MARKUP})

to this:

 BeautifulSoup(YOUR_MARKUP, "lxml")

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>